+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Reworded the "incorrect network version" message to hopefully be clearer [#367]
 ### Fixed
 - Starlight charger displaying incorrect/negative stored charge on servers